progs.h: leafnums array in edict_t needs to be int, not short. patch from Eric Wasylishen.
git-svn-id: svn://svn.code.sf.net/p/quakespasm/code/trunk/quakespasm@899 af15c1b1-3010-417e-b628-4374ebc0bcbd

+++ Quake/progs.h
@@ -42,7 +42,7 @@ typedef struct edict_s
 	link_t		area;			/* linked to a division node or leaf */
 
 	int		num_leafs;
-	short		leafnums[MAX_ENT_LEAFS];
+	int		leafnums[MAX_ENT_LEAFS];
 
 	entity_state_t	baseline;
 	unsigned char	alpha;			/* johnfitz -- hack to support alpha since it's not part of entvars_t */
